foreign keys type changed to 'long'

+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/AdvancedSlotSettings.cs
@@ -4,7 +4,7 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class AdvancedSlotSettings : Entity<long>
 {
-    public int AvailabilitySlotId { get; set; }
+    public long AvailabilitySlotId { get; set; }
     public ActivityType ActivityType { get; set; }
     public int Days { get; set; }
     public DateTimeOffset StartDate { get; set; }
