+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/AvailabilitySlot.cs
@@ -4,9 +4,9 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class AvailabilitySlot : AuditEntity<long>
 {
-    public int TeamId { get; set; }
-    public int LocationId { get; set; }
-    public int? AdvancedSlotSettingsId { get; set; }
+    public long TeamId { get; set; }
+    public long LocationId { get; set; }
+    public long? AdvancedSlotSettingsId { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string Link { get; set; } = string.Empty;
