+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/CalendarVisibleForTeam.cs
@@ -2,8 +2,8 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class CalendarVisibleForTeam : Entity<long>
 {
-    public int CalendarId { get; set; }
-    public int TeamId { get; set; }
+    public long CalendarId { get; set; }
+    public long TeamId { get; set; }
 
     public Calendar Calendar { get; set; }
     public Team Team { get; set; }
