+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/UserSlot.cs
@@ -2,8 +2,8 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class UserSlot : Entity<long>
 {
-    public int UserId { get; set; }
-    public int AvailabilitySlotId { get; set; }
+    public long UserId { get; set; }
+    public long AvailabilitySlotId { get; set; }
 
     public User User { get; set; }
     public AvailabilitySlot AvailabilitySlot { get; set; }
